move eventlistener out of react
